% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_screen_dims.R
\name{calculate_screen_dims}
\alias{calculate_screen_dims}
\title{Calculate effective screen dimensions}
\usage{
calculate_screen_dims(
  data,
  uas = TRUE,
  inner_height = NULL,
  inner_width = NULL,
  scroll_height = NULL,
  scroll_width = NULL,
  screen_height = NULL,
  screen_width = NULL
)
}
\arguments{
\item{data}{Any mouse trajectory dataset that contains inner and scroll dimensions of the screen.}

\item{uas}{Logical. If TRUE, use UAS defaults for columns and filtering.}

\item{inner_width, inner_height, scroll_width, scroll_height}{Column specs (string).
Inner dimensions correspond to the screen visible in the browser while scroll dimensions are of the physical monitor.}

\item{screen_width, screen_height}{Screen height and screen width. If it exists in the data and wish to
replace with the calculation, input the variable names.}
}
\value{
Dataset with (re)computed screen width and screen height columns.
}
\description{
Computes row-wise "effective" screen width and height as the maximum of:
screen width: max(inner_width, scroll_width) and screen height: max(inner_height, scroll_height).
}
\details{
If \code{uas = TRUE}, the function uses UAS default column names and defaults for
filtering columns, so you can call the function with only \code{data}.
}
\examples{
# screen_dims_data <- calculate_screen_dims(df) 

}
