% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_cases.R
\name{rm_cases}
\alias{rm_cases}
\title{Remove cases from a dataset}
\usage{
rm_cases(
  data,
  part_id = "mouseid",
  screen_id = "mt_id",
  column_rm,
  factor_rm,
  criteria
)
}
\arguments{
\item{data}{A data frame containing mouse movement data.}

\item{part_id}{Column name of the participant identifier.}

\item{screen_id}{Column name of the screen or worker identifier. One participant
can have multiple screen IDs.}

\item{column_rm}{Column name of the variable used for filtering.}

\item{factor_rm}{The value or category that should be removed.}

\item{criteria}{Integer indicating the removal level:
\describe{
\item{1}{Remove only the screen/worker ID where the criterion is met.}
\item{2}{Remove all data from the participant where the criterion is met.}
}}
}
\value{
A filtered data frame with the specified cases removed.
}
\description{
This function removes cases from a dataset based on a specified filter variable
and criterion. Filtering can be applied either at the screen (case) level or at
the participant level.
}
\details{
The function can be used to exclude cases where, for example:
\itemize{
\item the browser window was resized,
\item a touch device was used,
\item the response language was not English,
\item other exclusion criteria apply.
}
}
\examples{
rm_cases(data, column_rm = "userAgent_is_touch_capable", factor_rm = TRUE, criteria = 1)

}
