% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_part.R
\name{rm_part}
\alias{rm_part}
\title{Remove participants or screens from a dataset}
\usage{
rm_part(data, part_id = "mt_id", rm_vector)
}
\arguments{
\item{data}{A data frame containing mouse movement data.}

\item{part_id}{Column name of the participant or screen identifier.}

\item{rm_vector}{A vector of all IDs that need to be removed.}
}
\value{
A filtered data frame with the specified participants or screens removed.
}
\description{
This function removes participants or screens from a dataset based on their ID.
}
\examples{
new_df <- rm_part(df, "mt_id", part_vec)

}
