% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_cases.R
\name{sl_cases}
\alias{sl_cases}
\title{Selects cases from dataset}
\usage{
sl_cases(data, column_sl, factor_sl)
}
\arguments{
\item{data}{A data frame with the mouse movement data.}

\item{column_sl}{The column, which is used for the filter selection.}

\item{factor_sl}{The factor / category which is selected.}
}
\value{
A filtered data frame with removed cases.
}
\description{
This function selects the specified cases based on the column and the level, return the
new data frame.
}
\details{
Can be used to select cases with e.g.:
\itemize{
\item no touch device
\item mouse movement data
\item ...
}
}
\examples{
sl_cases(data, column_sl = "type", factor_sl = "mousemove")
}
