% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_cols.R
\name{standardize_colnames}
\alias{standardize_colnames}
\title{Standardize common column names to package conventions}
\usage{
standardize_colnames(data, verbose = TRUE)
}
\arguments{
\item{data}{A data frame / tibble.}

\item{verbose}{Logical. If TRUE, prints a small summary of renames. The default is TRUE.}
}
\value{
\code{data} with renamed columns.
}
\description{
This function renames frequently used variables to a standard naming scheme for mouse processing.
Rule examples (case-insensitive):
\itemize{
\item any name containing "worker" and "id"  -> "mt_id"
\item any name containing "time" and "stamp" -> "time_stamp"

}
}
\details{
The function ican be extended to add more rules to the \code{rules} table later.
}
\examples{
# df2 <- standardize_colnames(df, verbose = TRUE)

}
