% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_clicks.R
\name{include_clicks}
\alias{include_clicks}
\title{Include clicks to the preprocessed dataset}
\usage{
include_clicks(
  data_complete,
  data_clicks,
  screen_id = "mt_id",
  click_column = "timestamps",
  timestamps = "timestamps",
  initiation_time = "initiation_time"
)
}
\arguments{
\item{data_complete}{A data frame with the screen ID, timestamps, and repeated initiation time}

\item{data_clicks}{A data frame with the screen ID and timestamps of the clicks}

\item{screen_id}{The screen or worker ID, identifies the separated screens, one participant has multiple IDs. The screen ID should be identical for both data frames.}

\item{click_column}{The click column.}

\item{timestamps}{The timestamp column.}

\item{initiation_time}{The initiation time column, with the initiation time repeated for each screen.}
}
\value{
The original data frame with an additional click column.
}
\description{
This function takes the data, timestamps, initiation time, and click times, and adds a column indicating clicks using one-hot encoding.
Clicks are aligned to the closest available timestamp. As a result, the original click data frame and the resulting data frame may differ in row count.
}
\examples{
click_data <- include_clicks(df_complete, df_click, click_column = "timestamps")
}
