% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_class.R
\name{mouse_class}
\alias{mouse_class}
\title{Filter mouse movement cases by number of recorded data points}
\usage{
mouse_class(
  data,
  screen_id = "mt_id",
  count_var = "timestamps",
  max_cutoff = 50,
  min_cutoff = 10
)
}
\arguments{
\item{data}{A (sl_cases()) data frame with only mouse movement data.}

\item{screen_id}{The grouping variable, like a screen or worker ID.}

\item{count_var}{A column in the dataset used to count the cases per screen_id.}

\item{max_cutoff}{The cutoff value for mouse users.}

\item{min_cutoff}{the cutoff value for touch users, should not be the same as max_cutoff.}
}
\value{
An object of class \code{mouse_class}.
}
\description{
This function filters participant cases based on the number of recorded
mouse or touch movement data points per screen, using the timestamp variable.
By default, cases with more than 50 data points are classified as mouse users,
and cases with fewer than 10 data points are classified as touch device users.
Cases with 10–50 data points are considered ambiguous and can be manually
inspected (e.g., via trajectory plots).
}
\details{
If only \code{max_cutoff} is specified, ambiguous devices are removed and the
filtered dataset is returned.
If both \code{min_cutoff} and \code{max_cutoff} are specified, cases within the
specified range are returned for further inspection.
If only \code{min_cutoff} is specified, only ambiguous device cases are returned.
}
\examples{
df_mv <- sl_cases(df, column_sl = "type", factor_sl = "mousemove")
df_mv <- mouse_class(df_mv, max_cutoff = 50, min_cutoff = 10)
}
