% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_class.R
\name{mouse_class}
\alias{mouse_class}
\title{Classification of touch/mouse users.}
\usage{
mouse_class(
  data,
  sess_id = "workerId",
  count_var = "timeStamp",
  max_cutoff = FALSE,
  min_cutoff = FALSE
)
}
\arguments{
\item{data}{A (sl_cases()) data frame with only mouse movement data.}

\item{sess_id}{The grouping variable, like a session or worker ID.}

\item{count_var}{A column in the dataset used to count the cases per sess_id.}

\item{max_cutoff}{The cutoff value for mouse users.}

\item{min_cutoff}{the cutoff value for touch users, should not be the same as max_cutoff.}
}
\value{
An object of class \code{mouse_class}.
}
\description{
This function selects cases based on the amount of recorded data points per session
per participant. Default value is >50 for mouse users and <10 for touch devices. Cases between
10 and 50 data points can be manually inspected, e.g. by checking the trajectory plots.
}
\details{
Define only max_cutoff, when the questionable devices need to be removed, returns filtered dataset.
When min_cutoff is also specified, the cases in the specified range will be returned for further inspection.
When only min_cutoff is specified, the function returns the questionable devices only.
}
\examples{
df_mv <- sl_cases(df, column_sl = "type", factor_sl = "mousemove")
df_mv <- mouse_class(df_mv, max_cutoff = 50, min_cutoff = 10)
}
