% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_process_data.R
\name{mt_process_data}
\alias{mt_process_data}
\title{Process mouse-movement data with mousetrap}
\usage{
mt_process_data(
  data,
  nsteps,
  hover_threshold = 500,
  non_na_cols = c("timestamps", "xpos", "ypos"),
  id = "mt_id"
)
}
\arguments{
\item{data}{A data frame containing raw mouse-tracking data in long format,
structured for use with \code{mt_import_long()}.}

\item{nsteps}{An integer giving the number of steps to use for
time-normalization of trajectories.}

\item{hover_threshold}{A numeric threshold in milliseconds passed to
\code{mt_measures()} for hover detection. Defaults to \code{500}.}
}
\value{
A mousetrap object containing: imported raw trajectories,
time-normalized trajectories saved as \verb{tn_trajectories_<nsteps>},
derivative measures, summary measures saved as \verb{measures_<nsteps>}.

The time-normalized trajectories are stored under the name
\verb{tn_trajectories_<nsteps>}, and the measures are stored under the name
\verb{measures_<nsteps>}. The \code{data} object contains the derivative values.
}
\description{
Imports mouse-tracking data in the long format, time-normalizes trajectories, computes
derivatives, and calculates summary measures using the \code{mousetrap} package.
}
\examples{
processed_data <- mt_process_data(data = trajectories, nsteps = 101,hover_threshold = 500)


}
