% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_traj.R
\name{multiple_traj}
\alias{multiple_traj}
\title{Detect screen with multiple visits}
\usage{
multiple_traj(data, part_id, quest_id, screen_id = "mt_id", criteria)
}
\arguments{
\item{data}{A (sl_cases()) data frame with only mouse movement data.}

\item{part_id}{The ID of the participant.}

\item{quest_id}{The ID of the question.}

\item{screen_id}{The ID on screen level.}

\item{criteria}{Character string specifying how to handle multiple
trajectories per screen:
\itemize{
\item \code{"all"} – remove the complete participant if multiple trajectories are present.
\item \code{"first"} – keep only the first recorded trajectory per screen.
\item \code{"last"} – keep only the last recorded trajectory per screen.
\item \code{"long"} – keep only the longest trajectory per screen
(based on the number of recorded data points).
}}
}
\value{
Initial dataframe with an indicator column names "one_obs_per_screen"
}
\description{
Detecting screens with multiple visits and keep a selected screen trajectory
}
