% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_cases.R
\name{rm_cases}
\alias{rm_cases}
\title{Remove cases from dataset}
\usage{
rm_cases(
  data,
  part_id = "mouseid",
  sess_id = "workerId",
  column_rm,
  factor_rm,
  criteria
)
}
\arguments{
\item{data}{A data frame with the mouse movement data.}

\item{part_id}{The participant ID.}

\item{sess_id}{The session or worker ID, identifies the separated session, one participant has multiple IDs.}

\item{column_rm}{The column, which is used for the filter selection.}

\item{factor_rm}{The factor / category which should be removed.}

\item{criteria}{For 1: Remove only the session / worker Id, where the criteria are met.
For 2: Remove all data from the participant, where the criteria are met.}
}
\value{
A filtered data frame with removed cases.
}
\description{
This function takes the data and the specified participant ID, session ID, the selected column,
which entails the filter variable, the filter variable, the filter criteria and returns the filtered
data frame with the removed cases.
}
\details{
Can be used to remove cases where e.g.:
\itemize{
\item the window was resized
\item touch devices were used
\item the answer was not recorded in English
\item ...
}
}
\examples{
rm_cases(data, column_rm = "userAgent_is_touch_capable", factor_rm = TRUE, criteria = 1)
}
