% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_cases_time.R
\name{rm_cases_time}
\alias{rm_cases_time}
\title{Remove participants or cases from data set based on maximum time spent on a question}
\usage{
rm_cases_time(
  data,
  part_id = FALSE,
  screen_id = "workerId",
  time_col,
  max_time
)
}
\arguments{
\item{data}{A data frame with the mouse movement data.}

\item{part_id}{The participant ID.}

\item{screen_id}{The screen ID.}

\item{time_col}{The column name of the timestamp.}

\item{max_time}{The time cutoff specified in minutes.}
}
\value{
A filtered data frame with removed cases.
}
\description{
This function returns a filtered dataset based on the maximum time spend on a question on
either participant or screen level. If the function should only consider the cases, leave
the participant ID empty and only fill screen ID. If the function should consider the time on
participant level, fill in both participant and screen ID.
}
\details{
Can be used to remove cases where time exceeded a certain threshold per question, e.g. 7 minutes.
}
\examples{
rm_cases_time(df, part_id = "mouseid", screen_id = "workerId", time_col = "timeStamp", max_time = 7)
}
