% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_cases.R
\name{sl_cases}
\alias{sl_cases}
\title{Select cases from a dataset}
\usage{
sl_cases(data, column_sl, factor_sl)
}
\arguments{
\item{data}{A data frame containing mouse movement data.}

\item{column_sl}{Column name of the variable used for selection.}

\item{factor_sl}{The value or category to be selected.}
}
\value{
A filtered data frame containing only the selected cases.
}
\description{
This function selects cases from a dataset based on a specified variable and
value and returns a filtered data frame.
}
\details{
The function can be used to select cases such as:
\itemize{
\item non-touch device users,
\item observations containing mouse movement data,
\item other user-defined selection criteria.
}
}
\examples{
sl_cases(data, column_sl = "type", factor_sl = "mousemove")
}
